Move generate_classifier_data.py to scripts directory and update references
